Use 64 bits for durations
- Adds an opaque representation of nanosecond duration for type safety
  and correctness.

- Fixes #9.

+++ src/precision_clock.ml
@@ -4,7 +4,7 @@
 open Functional
 
 module Foreign = struct
-  external get_time : unit -> int = "nebula_precision_clock_get_time"
+  external get_time : unit -> int64 = "nebula_precision_clock_get_time"
 end
 
 let get_time =
